refactor(assistant): move app-builder stores out of memory/ into a new apps/ dir
