+++ assistant/src/__tests__/app-git-history.test.ts
@@ -13,7 +13,7 @@ import {
   getAppFileAtVersion,
   getAppHistory,
   restoreAppVersion,
-} from "../memory/app-git-service.js";
+} from "../apps/app-git-service.js";
 import { createApp, getAppDirPath, updateApp } from "../memory/app-store.js";
 
 describe("App Git History", () => {
