+++ assistant/src/__tests__/app-git-service.test.ts
@@ -4,7 +4,7 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { afterEach, beforeEach, describe, expect, test } from "bun:test";
 
-import { commitAppTurnChanges } from "../memory/app-git-service.js";
+import { commitAppTurnChanges } from "../apps/app-git-service.js";
 import {
   createApp,
   deleteApp,
