+++ assistant/src/__tests__/conversation-agent-loop-inference-profile.test.ts
@@ -215,7 +215,7 @@ mock.module("../memory/app-store.js", () => ({
   getAppsDir: () => "/tmp/apps",
 }));
 
-mock.module("../memory/app-git-service.js", () => ({
+mock.module("../apps/app-git-service.js", () => ({
   commitAppTurnChanges: () => Promise.resolve(),
 }));
