+++ assistant/src/__tests__/conversation-agent-loop.test.ts
@@ -374,7 +374,7 @@ mock.module("../memory/app-store.js", () => ({
   getAppsDir: () => "/tmp/apps",
 }));
 
-mock.module("../memory/app-git-service.js", () => ({
+mock.module("../apps/app-git-service.js", () => ({
   commitAppTurnChanges: () => Promise.resolve(),
 }));
