+++ assistant/src/__tests__/conversation-slash-unknown.test.ts
@@ -227,7 +227,7 @@ mock.module("../workspace/turn-commit.js", () => ({
   commitTurnChanges: async () => {},
 }));
 
-mock.module("../memory/app-git-service.js", () => ({
+mock.module("../apps/app-git-service.js", () => ({
   commitAppTurnChanges: async () => {},
 }));
