+++ assistant/src/apps/app-git-service.ts
@@ -20,9 +20,9 @@
 import { existsSync, readFileSync, writeFileSync } from "node:fs";
 import { join } from "node:path";
 
+import { getAppsDir, resolveAppDir } from "../memory/app-store.js";
 import { getLogger } from "../util/logger.js";
 import { getWorkspaceGitService } from "../workspace/git-service.js";
-import { getAppsDir, resolveAppDir } from "./app-store.js";
 
 const log = getLogger("app-git");
