+++ assistant/src/config/bundled-skills/app-builder/tools/app-refresh.ts
@@ -1,4 +1,4 @@
-import { setAppCommitMessage } from "../../../../memory/app-git-service.js";
+import { setAppCommitMessage } from "../../../../apps/app-git-service.js";
 import * as appStore from "../../../../memory/app-store.js";
 import { executeAppRefresh } from "../../../../tools/apps/executors.js";
 import {
