+++ assistant/src/config/bundled-skills/app-builder/tools/app-update.ts
@@ -1,4 +1,4 @@
-import { setAppCommitMessage } from "../../../../memory/app-git-service.js";
+import { setAppCommitMessage } from "../../../../apps/app-git-service.js";
 import * as appStore from "../../../../memory/app-store.js";
 import type { AppUpdateInput } from "../../../../tools/apps/executors.js";
 import { executeAppUpdate } from "../../../../tools/apps/executors.js";
