+++ assistant/src/runtime/routes/app-management-routes.ts
@@ -16,17 +16,18 @@ import { join } from "node:path";
 
 import { z } from "zod";
 
+import {
+  getAppDiff,
+  getAppHistory,
+  restoreAppVersion,
+} from "../../apps/app-git-service.js";
+import { createSharedAppLink } from "../../apps/shared-app-links-store.js";
 import { packageApp } from "../../bundler/app-bundler.js";
 import { compileApp } from "../../bundler/app-compiler.js";
 import { scanBundle } from "../../bundler/bundle-scanner.js";
 import type { SignatureJson } from "../../bundler/bundle-signer.js";
 import { verifyBundleSignature } from "../../bundler/signature-verifier.js";
 import { compareSemver } from "../../daemon/handlers/shared.js";
-import {
-  getAppDiff,
-  getAppHistory,
-  restoreAppVersion,
-} from "../../memory/app-git-service.js";
 import {
   type AppDefinition,
   createApp,
@@ -45,7 +46,6 @@ import {
   updateApp,
   updateAppRecord,
 } from "../../memory/app-store.js";
-import { createSharedAppLink } from "../../memory/shared-app-links-store.js";
 import { computeContentId } from "../../util/content-id.js";
 import { getLogger } from "../../util/logger.js";
 import { ACTOR_PRINCIPALS } from "../auth/route-policy.js";
