+++ assistant/src/runtime/routes/publish-routes.ts
@@ -13,6 +13,11 @@ import { join } from "node:path";
 import { v4 as uuid } from "uuid";
 import { z } from "zod";
 
+import {
+  getActivePublishedPageByAppId,
+  insertPublishedPage,
+  updatePublishedPage,
+} from "../../apps/published-pages-store.js";
 import { compileApp } from "../../bundler/app-compiler.js";
 import {
   getApp,
@@ -20,11 +25,6 @@ import {
   isMultifileApp,
   resolveEffectiveAppHtml,
 } from "../../memory/app-store.js";
-import {
-  getActivePublishedPageByAppId,
-  insertPublishedPage,
-  updatePublishedPage,
-} from "../../memory/published-pages-store.js";
 import { deployHtmlToVercel } from "../../services/vercel-deploy.js";
 import { credentialBroker } from "../../tools/credentials/broker.js";
 import { getLogger } from "../../util/logger.js";
